# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/pet": {
      "get": {
        "operationId": "PetController_getAllPets",
        "summary": "获取所有宠物列表",
        "description": "获取系统中所有可用的宠物列表。当 includeSelection=true 时，需要JWT认证，返回选中状态",
        "parameters": [
          {
            "name": "includeSelection",
            "required": false,
            "in": "query",
            "description": "是否包含选中状态（需要JWT认证）",
            "example": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PetListResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "获取选中状态需要JWT认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/pet/user/active": {
      "get": {
        "operationId": "PetController_getUserActivePet",
        "summary": "获取用户当前激活的宠物",
        "description": "获取当前用户正在使用的宠物",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPetDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "用户未选择宠物",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/pet/user/all": {
      "get": {
        "operationId": "PetController_getUserPets",
        "summary": "获取用户的所有宠物",
        "description": "获取用户选择过的所有宠物列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPetListResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/pet/profile/complete": {
      "get": {
        "operationId": "PetController_getCompleteProfile",
        "summary": "获取宠物完整档案（推荐）",
        "description": "一次性获取当前激活宠物的完整档案信息（包含简介、统计、成就、梦想）",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PetCompleteProfileResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "用户未选择宠物",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/pet/{id}/profile": {
      "get": {
        "operationId": "PetController_getPetProfile",
        "summary": "获取宠物完整档案",
        "description": "获取宠物的完整档案信息（包含扩展字段：简短描述、详细描述、性格标签、经典台词）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PetProfileDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "宠物不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ]
      }
    },
    "/api/pet/{id}/stats": {
      "get": {
        "operationId": "PetController_getPetStats",
        "summary": "获取宠物旅行统计",
        "description": "获取基于当前激活宠物的旅行统计数据（旅行天数、已旅行城市、累计步数、累计里程）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PetStatsDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "用户未拥有该宠物",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/pet/{id}/achievements": {
      "get": {
        "operationId": "PetController_getAchievements",
        "summary": "获取成就列表",
        "description": "获取成就列表（包含用户进度和状态）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AchievementListResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "用户未拥有该宠物",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/pet/{id}/dreams": {
      "get": {
        "operationId": "PetController_getDreams",
        "summary": "获取梦想进度列表",
        "description": "获取梦想进度列表（包含当前进度和完成状态）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DreamListResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "用户未拥有该宠物",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/pet/{id}": {
      "get": {
        "operationId": "PetController_getPetById",
        "summary": "获取宠物详情",
        "description": "根据ID获取宠物详细信息",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PetDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "宠物不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pet"
        ]
      }
    },
...
  },
  "info": {
    "title": "Claw Step API",
    "description": "Claw Step 应用后端 API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "ResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "错误信息",
            "example": "请求参数错误"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "PetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          },
          "name": {
            "type": "string",
            "description": "宠物名称",
            "example": "小狐狸"
          },
          "englishName": {
            "type": "string",
            "description": "宠物英文名",
            "example": "Fox"
          },
          "shortDescription": {
            "type": "string",
            "description": "简短描述（用于顶部显示）",
            "example": "爱薅羊毛，也爱犀利点评，每天都会给你一份毒舌版《避雷指南》。"
          },
          "longDescription": {
            "type": "string",
            "description": "详细描述（用于简介Tab）",
            "example": "爱薅羊毛，也爱犀利点评，每天都会给你一份毒舌版《避雷指南》。唯一的小问题是，偶尔需要担心它被店家拉黑。"
          },
          "imageUrl": {
            "type": "string",
            "description": "宠物头像URL",
            "example": "https://example.com/fox.jpg"
          },
          "personalityTags": {
            "description": "性格标签数组",
            "example": [
              "毒舌",
              "爱薅羊毛",
              "机智"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "classicLines": {
            "description": "经典台词数组",
            "example": [
              "\"这味道，不如把我扔进榨汁机。\"",
              "\"这杯咖啡卖三十？这不是拿智商当奶泡打吗？\""
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isSelected": {
            "type": "boolean",
            "description": "是否已选中（当 includeSelection=true 时返回）",
            "example": true
          }
        },
        "required": [
          "id",
          "name",
          "englishName",
          "imageUrl"
        ]
      },
      "PetListResponseDto": {
        "type": "object",
        "properties": {
          "pets": {
            "description": "宠物列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PetDto"
            }
          },
          "currentPetId": {
            "type": "string",
            "description": "当前选中的宠物ID（当 includeSelection=true 时返回）",
            "example": "uuid-string"
          }
        },
        "required": [
          "pets"
        ]
      },
      "UserPetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "关系ID",
            "example": "uuid-string"
          },
          "userId": {
            "type": "string",
            "description": "用户ID",
            "example": "uuid-string"
          },
          "petId": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否激活",
            "example": true
          },
          "selectedAt": {
            "format": "date-time",
            "type": "string",
            "description": "选择时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "pet": {
            "description": "宠物信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/PetDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "petId",
          "isActive",
          "selectedAt",
          "pet"
        ]
      },
      "UserPetListResponseDto": {
        "type": "object",
        "properties": {
          "pets": {
            "description": "用户宠物列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPetDto"
            }
          }
        },
        "required": [
          "pets"
        ]
      },
      "PetProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "宠物ID",
            "example": "uuid-string"
          },
          "name": {
            "type": "string",
            "description": "宠物名称",
            "example": "艾迪"
          },
          "englishName": {
            "type": "string",
            "description": "宠物英文名",
            "example": "Eddie"
          },
          "shortDescription": {
            "type": "string",
            "description": "简短描述（用于顶部显示）",
            "example": "爱薅羊毛，也爱犀利点评，每天都会给你一份毒舌版《避雷指南》。"
          },
          "longDescription": {
            "type": "string",
            "description": "详细描述（用于简介Tab）",
            "example": "爱薅羊毛，也爱犀利点评，每天都会给你一份毒舌版《避雷指南》。唯一的小问题是，偶尔需要担心它被店家拉黑。"
          },
          "imageUrl": {
            "type": "string",
            "description": "宠物头像URL",
            "example": "https://example.com/fox.jpg"
          },
          "personalityTags": {
            "description": "性格标签数组",
            "example": [
              "毒舌",
              "爱薅羊毛",
              "机智"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "classicLines": {
            "description": "经典台词数组",
            "example": [
              "\"这味道，不如把我扔进榨汁机。\"",
              "\"这杯咖啡卖三十？这不是拿智商当奶泡打吗？\""
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "englishName",
          "imageUrl",
          "personalityTags",
          "classicLines",
          "createdAt",
          "updatedAt"
        ]
      },
      "PetStatsDto": {
        "type": "object",
        "properties": {
          "travelDays": {
            "type": "number",
            "description": "旅行天数",
            "example": 87
          },
          "visitedCities": {
            "type": "number",
            "description": "已旅行城市数量",
            "example": 15
          },
          "totalSteps": {
            "type": "number",
            "description": "累计步数",
            "example": 50000
          },
          "totalDistance": {
            "type": "number",
            "description": "累计旅行里程（公里）",
            "example": 35
          }
        },
        "required": [
          "travelDays",
          "visitedCities",
          "totalSteps",
          "totalDistance"
        ]
      },
      "AchievementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "成就ID",
            "example": "uuid-string"
          },
          "name": {
            "type": "string",
            "description": "成就名称",
            "example": "日行万步"
          },
          "description": {
            "type": "string",
            "description": "成就描述",
            "example": "一天内行走10000步"
          },
          "iconUrl": {
            "type": "string",
            "description": "成就图标URL",
            "example": "https://example.com/icon.png"
          },
          "status": {
            "type": "string",
            "description": "成就状态",
            "example": "unlocked",
            "enum": [
              "locked",
              "unlocked",
              "in_progress"
            ]
          },
          "currentProgress": {
            "type": "number",
            "description": "当前进度值",
            "example": 10000
          },
          "targetValue": {
            "type": "number",
            "description": "目标值",
            "example": 10000
          },
          "unlockedAt": {
            "format": "date-time",
            "type": "string",
            "description": "解锁时间",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "status"
        ]
      },
      "AchievementListResponseDto": {
        "type": "object",
        "properties": {
          "achievements": {
            "description": "成就列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AchievementDto"
            }
          }
        },
        "required": [
          "achievements"
        ]
      },
      "DreamDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "梦想ID",
            "example": "uuid-string"
          },
          "name": {
            "type": "string",
            "description": "梦想名称",
            "example": "编写《环球薅羊毛圣经》"
          },
          "description": {
            "type": "string",
            "description": "梦想描述",
            "example": "正在为成为世界首席省钱顾问而努力！"
          },
          "icon": {
            "type": "string",
            "description": "图标（emoji或URL）",
            "example": "📖"
          },
          "currentProgress": {
            "type": "number",
            "description": "当前进度值",
            "example": 12
          },
          "targetValue": {
            "type": "number",
            "description": "目标值",
            "example": 100
          },
          "progressPercent": {
            "type": "number",
            "description": "进度百分比",
            "example": 0.12
          },
          "status": {
            "type": "string",
            "description": "状态",
            "example": "in_progress",
            "enum": [
              "in_progress",
              "completed"
            ]
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "description": "完成时间",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "currentProgress",
          "targetValue",
          "progressPercent",
          "status"
        ]
      },
      "DreamListResponseDto": {
        "type": "object",
        "properties": {
          "dreams": {
            "description": "梦想列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DreamDto"
            }
          }
        },
        "required": [
          "dreams"
        ]
      },
      "PetCompleteProfileResponseDto": {
        "type": "object",
        "properties": {
          "pet": {
            "description": "宠物信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/PetProfileDto"
              }
            ]
          },
          "stats": {
            "description": "旅行统计",
            "allOf": [
              {
                "$ref": "#/components/schemas/PetStatsDto"
              }
            ]
          },
          "achievements": {
            "description": "成就列表",
            "allOf": [
              {
                "$ref": "#/components/schemas/AchievementListResponseDto"
              }
            ]
          },
          "dreams": {
            "description": "梦想列表",
            "allOf": [
              {
                "$ref": "#/components/schemas/DreamListResponseDto"
              }
            ]
          }
        },
        "required": [
          "pet",
          "stats",
          "achievements",
          "dreams"
        ]
      },
...
    }
  }
}